Aula3/4: Add private in funcs and UIView to ViewController
